project/kafka-services/consumer-network/requirements.txt
pandas
networkx
psycopg2-binary
confluent-kafka
